--- a/src/sentry/options/defaults.py
+++ b/src/sentry/options/defaults.py
@@ -1227,12 +1227,12 @@ register(
 )
 register(
     "performance.issues.n_plus_one_db.duration_threshold",
-    default=100.0,
+    default=90.0,
     flags=FLAG_AUTOMATOR_MODIFIABLE,
 )
 register(
     "performance.issues.slow_db_query.duration_threshold",
-    default=1000.0,  # ms
+    default=900.0,  # ms
     flags=FLAG_AUTOMATOR_MODIFIABLE,
 )
 register(
@@ -1257,7 +1257,7 @@ register(
 )
 register(
     "performance.issues.consecutive_http.max_duration_between_spans",
-    default=500,
+    default=500,  # ms
     flags=FLAG_AUTOMATOR_MODIFIABLE,
 )
 register(
@@ -1267,7 +1267,12 @@ register(
 )
 register(
     "performance.issues.consecutive_http.span_duration_threshold",
-    default=1000,
+    default=500,  # ms
+    flags=FLAG_AUTOMATOR_MODIFIABLE,
+)
+register(
+    "performance.issues.consecutive_http.min_time_saved_threshold",
+    default=2000,  # ms
     flags=FLAG_AUTOMATOR_MODIFIABLE,
 )
 register(