--- a/src/sentry/utils/performance_issues/detectors/consecutive_http_detector.py
+++ b/src/sentry/utils/performance_issues/detectors/consecutive_http_detector.py
@@ -50,6 +50,10 @@ class ConsecutiveHTTPSpanDetector(PerformanceDetector):
         if not span_id or not self._is_eligible_http_span(span):
             return
 
+        span_duration = get_span_duration(span).total_seconds() * 1000
+        if span_duration < self.settings.get("span_duration_threshold"):
+            return
+
         if self._overlaps_last_span(span):
             self._validate_and_store_performance_problem()
             self._reset_variables()
@@ -63,11 +67,12 @@ class ConsecutiveHTTPSpanDetector(PerformanceDetector):
         exceeds_count_threshold = len(self.consecutive_http_spans) >= self.settings.get(
             "consecutive_count_threshold"
         )
-        exceeds_span_duration_threshold = all(
-            get_span_duration(span).total_seconds() * 1000
-            > self.settings.get("span_duration_threshold")
-            for span in self.consecutive_http_spans
-        )
+
+        exceeds_min_time_saved_duration = False
+        if self.consecutive_http_spans:
+            exceeds_min_time_saved_duration = self._calculate_time_saved() >= self.settings.get(
+                "min_time_saved"
+            )
 
         subceeds_duration_between_spans_threshold = all(
             get_duration_between_spans(
@@ -79,10 +84,16 @@ class ConsecutiveHTTPSpanDetector(PerformanceDetector):
 
         if (
             exceeds_count_threshold
-            and exceeds_span_duration_threshold
             and subceeds_duration_between_spans_threshold
+            and exceeds_min_time_saved_duration
         ):
             self._store_performance_problem()
+
+    def _calculate_time_saved(self) -> float:
+        total_time = get_total_span_duration(self.consecutive_http_spans)
+        max_span_duration = get_max_span_duration(self.consecutive_http_spans)
+
+        return total_time - max_span_duration
 
     def _store_performance_problem(self) -> None:
         fingerprint = self._fingerprint()
@@ -166,73 +177,3 @@ class ConsecutiveHTTPSpanDetector(PerformanceDetector):
 
     def is_creation_allowed_for_project(self, project: Project) -> bool:
         return self.settings["detection_enabled"]
-
-
-class ConsecutiveHTTPSpanDetectorExtended(ConsecutiveHTTPSpanDetector):
-    """
-    Detector goals:
-    - Extend Consecutive HTTP Span Detector to mimic detection using thresholds from
-    - Consecutive DB Queries Detector.
-    """
-
-    type = DetectorType.CONSECUTIVE_HTTP_OP_EXTENDED
-    settings_key = DetectorType.CONSECUTIVE_HTTP_OP_EXTENDED
-
-    def visit_span(self, span: Span) -> None:
-        if is_event_from_browser_javascript_sdk(self.event()):
-            return
-
-        span_id = span.get("span_id", None)
-
-        if not span_id or not self._is_eligible_http_span(span):
-            return
-
-        span_duration = get_span_duration(span).total_seconds() * 1000
-        if span_duration < self.settings.get("span_duration_threshold"):
-            return
-
-        if self._overlaps_last_span(span):
-            self._validate_and_store_performance_problem()
-            self._reset_variables()
-
-        self._add_problem_span(span)
-
-    def _validate_and_store_performance_problem(self):
-        exceeds_count_threshold = len(self.consecutive_http_spans) >= self.settings.get(
-            "consecutive_count_threshold"
-        )
-
-        exceeds_min_time_saved_duration = False
-        if self.consecutive_http_spans:
-            exceeds_min_time_saved_duration = self._calculate_time_saved() >= self.settings.get(
-                "min_time_saved"
-            )
-
-        subceeds_duration_between_spans_threshold = all(
-            get_duration_between_spans(
-                self.consecutive_http_spans[idx - 1], self.consecutive_http_spans[idx]
-            )
-            < self.settings.get("max_duration_between_spans")
-            for idx in range(1, len(self.consecutive_http_spans))
-        )
-
-        if (
-            exceeds_count_threshold
-            and subceeds_duration_between_spans_threshold
-            and exceeds_min_time_saved_duration
-        ):
-            self._store_performance_problem()
-
-    def _calculate_time_saved(self) -> float:
-        total_time = get_total_span_duration(self.consecutive_http_spans)
-        max_span_duration = get_max_span_duration(self.consecutive_http_spans)
-
-        return total_time - max_span_duration
-
-    def is_creation_allowed_for_organization(self, organization: Organization) -> bool:
-        # Only collecting metrics.
-        return False
-
-    def is_creation_allowed_for_project(self, project: Project) -> bool:
-        # Only collecting metrics.
-        return False