For project: [Detector Threshold Configurations](https://www.notion.so/sentry/Detector-Threshold-Configuration-f8cb07e7cceb42388cedb09ea05fc116)
Lowered defaults values and effects: [link](https://www.notion.so/sentry/Detector-Dry-Run-Results-40dc7a3e4d8b4b9ea8da90608fe54747)

- We are GA'ing the project with lowered default values for thresholds and adding new thresholds.
- Changed default threshold for Consecutive HTTP to 500ms to match prod sentry_option value: [query](https://redash.getsentry.net/queries/4546/source)
- Added lowered defaults for N+1 DB duration (iterating down to 50ms), Slow DB Queries duration(iterating down to 500ms), Consecutive http span duration (iterating down to 500ms)
- Updated Consecutive HTTP Detector to add new total min time saved threshold with a default of 2000s.
- Updated Consecutive HTTP Detector  tests to ensure new total duration threshold is respected.
- Removed Extended Consecutive HTTP Detector  and tests.